+++ src/main/java/com/heartsave/todaktodak_api/auth/dto/request/NicknameCheckRequest.java
@@ -1,4 +1,4 @@
-package com.heartsave.todaktodak_api.auth.dto.response;
+package com.heartsave.todaktodak_api.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
